Python_Lections/Projects/django_taskmanager/django_taskmanager/requirements.txt
django>=4.2,<5.1
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3
